# Optimize performance when gathering content pages