@jasonvarga 2024-06-24
Did you close this because `url()` is faster in Statamic 5?